css/css-flexbox/parsing/justify-content-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Flexible Box Layout Module Level 1: parsing justify-content with invalid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-flexbox/#propdef-justify-content">
<meta name="assert" content="justify-content supports only the grammar 'flex-start | flex-end | center | space-between | space-around'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("justify-content", "auto");

test_invalid_value("justify-content", "space-around center"); // Accepted by Edge/WebKit.
</script>
</body>
</html>
